Bug fixes
Fixed bug in /fix shop that caused shop items to move to the misc category.

Fixed bug in IRC command 'list event'

Blocked template entitities and invisible enemies from spawning via the bot.

Reset regions recorded on the server (in the Botman mod) are now read and used as the true record of reset regions so there should not be any more extra reset regions that don't belong.

When the server is rebooting and is busy resetting regions, the progress is echoed to IRC.