Gimme zombies fix and a new debug log
The command /fix gimme now empties the table of zombies and rebuilds it.  It used to do this and was changed but I found that it must be completely refreshed or the entity id's become wrong.  Entity ID is a key field.

There is a new debug log in the daily logs.  When debugging is written to it there will be a log so it will only generate a log on days when it is used.  The bot is currently logging all 6 places in the bot's code where it switches from api mode to telnet mode so I can track and fix that bug.