Fixes and tweaks
Added gimme prizes and gimme zombies to the backup and added the ability to only import zombies or prizes from the backup.

Changed some stuff in the reconnect timer so that the bot won't switch from API mode to telnet mode so quickly and also blocked it from doing so if the bot is not connected or the server is starting up.