Bug fixes
Fixed a timestamp bug.  The bot was getting the correct server timestamp but then was adding to it wrongly.  Now it just grabs it from the server a lot and doesn't try to calculate it.

Added an extra check before initialising a player as new.  If the bot started less than 2 minutes ago, it will try to load the player record and will exit the rest of the code.  That should give the bot time to get the player record loaded before it tries to check for it.  If the player record doesn't exist, the bot will carry on an initialise the player as new after the bot has been running longer than 2 minutes.

Resetting the bot now also clears all reset regions from the Botman mod (live map).

Template and invisible zombies should be automatically removed from gimme zombies and otherEntities lists.