Addition to the command /trader protect
Here is the new command help

/trader protect/unprotect/remove {named area}
/trader add/trader del {named area}
/trader protect/unprotect north/south/east/west/ne/nw/se/sw length/long {number} width/wide {number}
After marking out a named area with the /mark command, you can add or remove trader protection on it.
When used with a compass direction you do not need to mark out the area first. Your current position is used instead.
Also when a compass direction is used, do not give the area a name as it is not recorded. You can repeat the command by simply typing /

The new part is the ability to specify a compass direction and width and length.  If you only supply width or length the missing value gets the value you did provide eg. width.  The old version of this command is still there and unchanged.  The new variant works differently, not using markings or a name.