Some bug fixes.
Small adjustment to the bot's parsing of reset regions in the Botman mod.  The mod and bot should be in sync.

Fixed bug in /give player {name} item.  It was getting confused if you used the steam ID or game ID and trying to give the player crazy amounts of the item.  I am pretty sure the game never allowed that to actually happen anyway.

Fixed missing help line for /set server vote reward.
Fixed bug in /set server vote reward that prevented it accepting the item name.  It kept saying it wasn't valid.

Fixed missing help line in /track command.