bug fixes and new commands
Minor tweak to code that is supposed to prevent the bot rebooting the server too soon after rebooting the server.  This is to allow plenty of time for the bot to get the correct server uptime and hopefully prevent it doing those negative uptime reboots.

Added new command to stop the bot telling players 'This command is disabled'.  Use this if you have another manager such as CSMM handling one or more commands that are also bot commands but you have disabled them in the bot and have both the bot and the other manager using the same command prefix (usually / ).
/show (or /hide) disabled command

Tweaked the way player group permissions are applied so that individual player overrides are possible for max waypoints, max bases, max protected bases, and reserved slot usage.

Fixed a display bug in The Lounge (irc) when the bot displays steam id. String formatting was somehow causing the last digit of the steam id in the who commands to change.  To be sure I removed the string formatting in a few more places in case the same issue could occur there.

Fixed a typo in the lounge (irc) command help that prevented some of the help displaying.

Completed a major overhaul of the main lounge (irc) help so that now all lounge commands are in individual functions including their command help.  This is in keeping with how I have organised the chat command code in the bot.  It is mainly for my benefit and makes code maintenance and just finding the relevant code a whole lot easier. Along with this, I added more command help for lounge commands and improved the command help in others.

Some minor tweaks to the bot's startup code to fix several issues, mainly with the auto-generated command help and also with suppressing those rogue reboots.

Added new lounge command 'list all items' which currently only works when the bot is in API mode.  What it does is it creates an items.txt file in the same location as the generated command help (in the temp subfolder of the daily logs folder). It lists all items known to your server including all items added by mods. Since it uses the console command li to generate, it is currently limited to just the item names used to spawn the items and does not include the name that players see. Also it is not sorted. The list is in the same order that the console command li produces.

I have removed support code for the botman.ini file which isn't really used for anything.

The in-game command /info {player} when used by a regular player will always list the player's own info no matter what they type after info.  It did this anyway but players were having issues with it.