fixes
Increased the timer in restart timer for how long the bot waits for a response from Alloc's web API before deciding that the API is not working to restart the bot.

Also made player group settings override a player's personal max waypoints, bases and protected bases unless they are not in a group.