Fixes and new commands
New command /play as player

Works like /test as player but lasts for 1 year or until the command '/restore admin' is used. While the new command is in effect, the admin player still has access to all the admin commands that their admin level allowed. As far as the server is concerned they are not admin but the bot remembers. Due to this change, if you want to remove an admin you must use the /admin remove command. The bot will automatically add admins via the console command admin list but it will not automatically remove them from its own internal admin list if they are removed from the server's admin list.

There currently is an issue with the /play as player command.  When used in game it is not able to remove the admin from the admin list if they are on the server at the time.  If it is used via The Lounge while the admin is not on the server, it works properly.

The bot will not arrest or otherwise punish an admin while using the /play as player feature since it knows they are still an admin so you can have admin items on you etc.

New command /reset chunk

/reset chunk
/reset chunk wide {number}
/reset chunk wide {number} player {player}
/reset chunk {player}
Reset chunks at your position or where another player is standing.
When used with a player name, the player must be on the server.
When used on yourself, you must be in-game.
ACL: 0 restricted to server owners, in-game and IRC

You can finally reset chunks using a chat command rather than needing to resort to the console command. The width setting means you don't need to mark a start and end point. It can be used via The Lounge but only when a player is specified and the player must be on the server at the time.

Player lookups are now aware of the Steam_ and XBL_ prefix so you no longer need to strip those out when doing lookups on steam id or xbox id.

Fixed bug in /bases command that prevented it listing the nearby bases.

The command /save {name}, used for saving a custom prefab now uses the Botman mod's bm-pexport console command instead of the Pimps own console command. This fixes the issue where freshly saved prefabs are unknown to the server until after the next reboot. Now those prefabs are available to spawn immediately.