Bug fixes and update
Updated the german code branch to bring it up to date with the a20 code branch.

Fixed bug in commands that remove an admin.

Fixed bug in command /location allow/disallow base

When the shop is disabled, typing /cash will tell players that the shop is closed.  Note that it is never closed for admins.

The command /reset timers now also resets the pack cooldown.