Bug fix
There was a bug in the donor expiry code if a donor was recorded using their EOS id but was in the players table using their steam id.  Also I moved the code to a dedicated function and call it in the 1 hour timer instead of at the end of the day.