Bug fixes
Fixed issue with restoring bot backups.  It needed a short delay added between restoring the data in the database and reloading the data from the database.

Fixed a bug in /setup map.  An earlier code change somehow got lost and the console commands sent to set map permissions were missing a required parameter.