Fixed bug in lounge chat
The help command done in the Lounge was failing to display help lines that start with a /  This was because the latest Mudlet was treating those as IRC commands.  Fixed by simply adding a space at the start of every line.  The space gets lost later but stops Mudlet messing it up.