Small fix for lounge chat bug
Recent changes to replace Server: with the bot name in the lounge worked a bit too well.  I have dialed it back a bit.

Also there is a new v1 code branch to be more in line with the Pimps naming.  The r1 code branch will auto switch bots to the V1 code branch in the next update coming soon.