Fixes and tweaks
The bot now auto reconnects to irc (The Lounge) every 2 minutes.  This will fix the occasional issue where the bot falls out of the channels.
Adjusted the behaviour of the bot if it discovers that the telnet password is wrong.  It no longer spams connection attempts and it will report to The Lounge that the telnet password is wrong.
Also there is new behaviour when the bot can't get a login prompt from telnet and some new status messages in The Lounge.
There is a very small fix in the API's admin list parser.  More testing and work is needed however as the bot still has occasional issues with the admin list in API mode.