New safe mode
Added safe mode to bot startup so it should not freeze if the server is spamming WRN or ERR.  It will also enter safe mode if it detects excessive spam at any time.  It will then briefly and slowly monitor the server at intervals and will only exit safe mode if the spam falls below a threshold.