important bugfix
On rare occasions (usually) the bot fails to see the telnet line saying the game has started.  It sets a flag for that and if that flag remains unset, teleport commands and other commands stop working (because the commands get queued and not sent).  Fixed by also setting the flag when a player joins since they can't do that when the server hasn't finished starting up.

Also improved the command help some more and started work on a new live bot status data table to allow for speedier investigations in the future when a bot isn't working correctly.