New group command and adjustments
Removed support for Alloc's mod older than webserver version 45.

Added new command to player groups to allow setting non-group stuff.

/no group set {setting number from list} value {a value}

When adding a player to a group you can now set an expiry and optional second group to move the player to when their group membership expires.

A command to set the expiry and second group for players who are already group members will be added soon but you use this command to update them as well.